code: hook idea
My thinking is to create a number of "hooks":
- beforeTransition
- afterTransition
- beforeNextTransition

When a route transition happens, two data items are calculated: the
current route and the route to move to:
- First, the `currentRoute.beforeNextTransition` is run. This allows the
current route the oportunity to "clean up" before the transitino
happens. Also, if a rejected promise is returned then the route
transition is aborted.
- `upcomingRoute.beforeTransition` is run. Same thing as before with the
rejected promise.
- The url change happens.
- `upcomingRoute.afterTransition` is run. Same thing as before with the
rejected promise.